Add files via upload
There are now 25 lines of numbers, the right most number is usually even for some reason and because of the process the lines are different lengths. they can be read as a exponent if you wish. Note run time is long because of seed that is made being huge, but the calls can be made larger by someone in more legitimist positioning with knowledge of how to compile with larger calls