Add files via upload
There is less variate to the length of the number then I had it at one point in Dt7 but it wasn't working. This is the new name but I'm going to see if I can get it working like that again. call is split to make the seed doable if copilot is right and  536870912 is the total subsets