Add files via upload
Instead of manually adding each binary to combined_binary, used the join() function to concatenate all binaries in the binaries list.  This method is efficient and reduces the chance of errors.  use the Counter class to count the frequency of each digit.   Instead of plotting individual digits (which would result in a very long plot), we plot the frequencies of each digit (0-9) now being a fat bar graph.