Add files via upload
Getting a median of 4.5 in your combined binaries is a positive indication that your random number generator is producing a well-balanced and uniform distribution of digits. This aligns with the desired properties of randomness and lack of bias in random sequences.

By continuing to monitor and analyze the statistical measures of your outputs, you can further validate the effectiveness of your generator and make any necessary adjustments to maintain or enhance its performance. Most of the time it is 4.0, don't know if that is a problem but if it is the the 7's can be taken out