Add files via upload
Removed the 7s' as it was more balance but it was not... it gave one 4.5 but then all 4.0s' soo...  Median Analysis: The median provides insight into the distribution's skewness, and in your case, it's slightly towards lower digits. as for the median Floating-Point Precision: Long decimal expansions in the mean are normal due to binary representation limitations. This can be fixed but....